Guessing a few more resources
